add alerts

+++ src/app/components/cart-app.component.ts
@@ -4,6 +4,8 @@ import { CartItem } from '../models/cart-item';
 import { NavbarComponent } from './navbar/navbar.component';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../services/sharing-data.service';
+import Swal from 'sweetalert2';
+
 @Component({
   selector: 'cart-app',
   standalone: true,
@@ -48,22 +50,45 @@ export class CartAppComponent implements OnInit {
       this.calculateTotal();
       this.saveSession();
       this.router.navigate(['/cart'], { state: { items: this.items, total: this.total } })
+      Swal.fire({
+        title: "Shopping Cart",
+        text: "Nuevo producto agregado al carro!",
+        icon: "success"
+      });
     });
   }
 
   onDeleteCart(): void {
     this.sharingDataService.idProductEventEmmiter.subscribe(id => {
       console.log(id + ' se ha ejecutado el evento idProductEventEmmiter');
-      this.items = this.items.filter(item => item.product.id !== id);
-      if (this.items.length == 0) {
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
+      Swal.fire({
+        title: "¿Está seguro que desea eliminar?",
+        text: "Cuidado!, el item se eliminará del carro de compras",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "¡Si, eliminar!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter(item => item.product.id !== id);
+          if (this.items.length == 0) {
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
 
-      this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
-        this.router.navigate(['/cart'], { state: { items: this.items, total: this.total } });
+          this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+            this.router.navigate(['/cart'], { state: { items: this.items, total: this.total } });
+          });
+
+          Swal.fire({
+            title: "¡Eliminado!",
+            text: "Producto eliminado del carro de compras",
+            icon: "success"
+          });
+        }
       });
     });
   }
